+++ packages/react/src/Form/MultiStep/index.ts
@@ -1,0 +1,3 @@
+export * from './FormStep';
+export * from './FormStepContext';
+export * from './FormStepHeader';