+++ packages/react/vite.config.js
@@ -51,13 +51,13 @@ export default defineConfig({
       // make sure to externalize deps that shouldn't be bundled
       external: [
         '@obosbbl/grunnmuren-icons',
-        '@react-hook/merged-ref',
+        'react-merge-refs',
+        '@react-hookz/web',
         'clsx',
         'react',
         'react-dom',
         'react/jsx-runtime',
         'react-collapsed',
-        'react-use',
       ],
     },
   },